requirements-dev.txt
-r requirements.txt
tox==4.16.0
pytest==8.2.2
pytest-asyncio==0.23.7
pytest-xdist==3.6.1
pytest-randomly==3.15.0
factory-boy==3.3.0
pytest-factoryboy==2.7.0
fakeredis==2.23.3
mypy==1.10.1
ruff==0.5.1
coverage==7.6.0
pytest-cov==5.0.0